backend/requirements.txt
Flask==2.3.2
lxml==4.9.3
python-dotenv==1.0.0
pytz==2025.2
python-dateutil==2.8.2
pytest==7.4.0
